// apps/ui/src/app/client/model/contentDto.ts
/**
 * LMStudio Wrapper API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ContentDto { 
    /**
     * The reasoning text from the model.
     */
    text: string;
    /**
     * The type of the reasoning text. Always `reasoning_text`.
     */
    type: ContentDto.TypeEnum;
}
export namespace ContentDto {
    export const TypeEnum = {
        ReasoningText: 'reasoning_text'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}